replace duration integer w float

+++ app/graphql/mutations/review_create.rb
@@ -12,7 +12,7 @@ module Mutations
     argument :safety_meter, Integer, required: true
     argument :gender, String, required: false
     argument :hygiene, Integer, required: false
-    argument :duration, Integer, required: false
+    argument :duration, Float, required: false
     argument :condoms, String, required: false
     argument :punctuality, String, required: false
 
@@ -61,15 +61,15 @@ module Mutations
       end
 
       @review = Review.create!(
-        client: @client, 
-        user: @user, 
-        body: body, 
-        rating: rating, 
-        size: size, 
-        payment: payment, 
-        vibe: vibe, 
-        date_again: date_again, 
-        safety_meter: safety_meter, 
+        client: @client,
+        user: @user,
+        body: body,
+        rating: rating,
+        size: size,
+        payment: payment,
+        vibe: vibe,
+        date_again: date_again,
+        safety_meter: safety_meter,
         gender: gender,
         hygiene: hygiene,
         punctuality: punctuality,
